pyproject.toml
[project]
name = "law-agent"
version = "0.1.0"
description = "Legal Review Agent - Multi-agent system for contract analysis, drafting, and advisory"
requires-python = ">=3.11"
dependencies = [
    # Agent Framework
    "langgraph>=1.0.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-anthropic>=0.3.0",

    # LLM Router
    "litellm>=1.0.0",

    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",

    # Database
    "asyncpg>=0.30.0",
    "pgvector>=0.3.0",

    # Document Parsing
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",

    # Search & Matching
    "rapidfuzz>=3.9.0",

    # Task Queue
    "celery[redis]>=5.4.0",
    "redis>=5.0.0",

    # Config & Utils
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "httpx>=0.27.0",

    # Embedding
    "openai>=1.50.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
